requirements.txt
PySide6>=6.6.0
python-docx>=1.1.0
